Index page style changes & @media queries
Remove the Bootstrap theme, to get the "vanilla" Bootstrap button style
and add additional styles for getting started buttons to be styled in
alignment with the overall theme.
Remove the Bootstrap theme, to get the "vanilla" Bootstrap button style
and add additional styles for getting started buttons to be styled in
alignment with the overall theme.
Site now behaves much better on portrait and landscape orientations on
mobile devices. The changes:
- Image (GIF) on index page not visible on mobile devices
	* This makes the navigation menu accessible again
- Getting started widths, margins, paddings and font sizes adjusted
  for smaller screen sizes.