Finishing work on the site
Added a footer include and some basic styling. Also changed some text
on the about page.
Fix various content bits on all the pages
Fixed various content pieces on all pages, mostly in style.
Finish work on footer
Finishing the contents of the footer and its styling.
Finishing the contents of the footer and its styling.