Add active designation on navigation links
Added JS code to highlight the "active" link on a given page.
Also added some overloads for hero buttons on main page for mobile
devices, mostly making them smaller and with less padding and smaller
margins.
Adding another menu item on the dotnet.github.io main page that leads
to the list of links. Added another section on the main page that lists
out the links to the sites. Some CSS modifications were also needed
since the top menu was making the link hard to see.
The page contains a slightly more detailed overview of all the
"moving pieces" in the new .NET ecosystem of 2015. Also add a nice
image that represents that ecosystem.