Move .NET Core site to main site
This commit moves the .NET Core site from the /core/
subdirectory to the root of the site. This makes the
dotnet.github.io site effectivelly the .NET Core site.
Projects listing, projects code and others have been removed.
This change includes all of the changes in the getting started
that happened recently.