Change getting started instructions & minor fixes
Change getting started instructions to use the new .NET CLI toolchain. Also,
add Docker instructions to the getting started page. Change the instructions
to use the Hello World sample from https://github.com/dotnet/core/ samples
so we don't have to upadte samples in 3 different places.
Fixes to CSS for links as well as to content to point to the new documentation folder and
API references.