Getting the site ready for RTM
This update will get the site ready for the RTM release. Since we have
the new docs site as well as the new website, it will have only one
page that will point people to API refs, docs and the getting started
content.

Also added redirects from /docs/ and /api/ and /support/ to the new
website.